fix circular imports and enum type
