--- cstar/io/__init__.py
@@ -1,43 +1,0 @@
-from cstar.io.retriever import (
-    LocalFileRetriever,
-    RemoteBinaryFileRetriever,
-    RemoteRepositoryRetriever,
-    RemoteTextFileRetriever,
-    Retriever,
-)
-from cstar.io.source_data import SourceClassification, SourceData, SourceDataCollection
-from cstar.io.staged_data import (
-    StagedData,
-    StagedDataCollection,
-    StagedFile,
-    StagedRepository,
-)
-from cstar.io.stager import (
-    LocalBinaryFileStager,
-    LocalTextFileStager,
-    RemoteBinaryFileStager,
-    RemoteRepositoryStager,
-    RemoteTextFileStager,
-    Stager,
-)
-
-__all__ = [
-    "StagedData",
-    "StagedFile",
-    "StagedDataCollection",
-    "StagedRepository",
-    "LocalBinaryFileStager",
-    "LocalTextFileStager",
-    "RemoteBinaryFileStager",
-    "RemoteRepositoryStager",
-    "RemoteTextFileStager",
-    "Stager",
-    "SourceData",
-    "SourceClassification",
-    "SourceDataCollection",
-    "Retriever",
-    "RemoteBinaryFileRetriever",
-    "RemoteTextFileRetriever",
-    "LocalFileRetriever",
-    "RemoteRepositoryRetriever",
-]
